Add note in Preview explaining Recomposition nuances of remember & state
- `remember` with `mutableStateOf` seems to cause a slightly different
pattern of recomposition that's probably tangentially related to "Strong
Skipping". `remember` helps notify Compose that updates are needed IN
THE Composable using `remember` based on the state's referential equality
and propagates changes down to its children! Without `remember` and
`mutableStateOf`, the stateless CreateReportView in the Preview
recomposes LESS since the Preview schedules updates ONLY IF the report
is structurally unequal. This equality-recomposition pattern is probably
noticeable due to `report.copy()` usage in event handlers that update
the `report` var. The relatively simple to recompose children of the
stateless CreateReportView still recompose as needed.