Add Screen-Level Stateful CreateReportView Composable
- Uses ViewModelCreateReport to pass down the created Report, the
text to display in the screen's header, and the data for the
HealthPractice. Additionally, callbacks are passed to update the
ViewModel state and submit the new Report via Composable UI interaction.
- Employee and Location data is not passed down since they
will be selected in other screens, then the selected data will be passed
back to this screen. Consequently, it may be simpler in the future to
provide only select properties of the ViewModel's Report.