Add artificial delay to make loading indicator more obvious
- Loading indicator is currently a linear indicator that is unfortunately
pretty subtle. Adding the delay gives the user more time to perceive it
while the UI gets the data and renders it. Luckily, thanks to the flow's
coroutine, it DOESN'T degrade the UX by freezing the UI. With more time
to perceive the loading indicator, the user can better understand the UI
ISN'T frozen and is indeed still working.
- Thanks to the delay, ViewModel tests need to use the UnconfinedTestDispatcher's
scheduler to bypass the delay. After doing so, the test can remove the
loading observer for tests to run and assert as expected.