Setup CreateReport ComposeView in Fragment + drop old Android View setup
- CreateReportView Composable setup involves taking the Fragment-scoped
ViewModel as input and observing the ViewModel from the Fragment to properly
setup the Composable screen. Without the Fragment observer, `arguments`
passed into the Fragment at instantiation cannot be received by the Composable
through the ViewModel to setup the Header Text composable
- Old Android Views are commented out for now EXCEPT for the AppBar Menu,
ProgressIndicator, and Snackbar, which can be adapted to Compose through
Scaffolds and scoping wrappers over time.